split credentials.tsx into focused modules
