+++ src/setup/config/model-selection.ts
@@ -1,0 +1,84 @@
+import {
+  getDefaultModelId,
+  getProviderModelOptions,
+  isValidModelId,
+  normalizeModelId,
+  OPENWIKI_MODEL_ID_ENV_KEY,
+  type OpenWikiProvider,
+} from "../../config/constants.js";
+
+type ModelSelectionOption =
+  | {
+      id: string;
+      kind: "preset";
+      label: string;
+    }
+  | {
+      kind: "custom";
+    };
+
+export function getModelSetupDetail(
+  modelIdOverride: string | null,
+  provider: OpenWikiProvider,
+): string {
+  if (modelIdOverride) {
+    return `using ${modelIdOverride} for this run`;
+  }
+
+  if (process.env[OPENWIKI_MODEL_ID_ENV_KEY]) {
+    return process.env[OPENWIKI_MODEL_ID_ENV_KEY] ?? "";
+  }
+
+  return `default ${getDefaultModelId(provider)}`;
+}
+
+export function getModelSelectionOptions(
+  provider: OpenWikiProvider,
+): ModelSelectionOption[] {
+  return [
+    ...getProviderModelOptions(provider).map((model) => ({
+      id: model.id,
+      kind: "preset" as const,
+      label: model.label,
+    })),
+    { kind: "custom" },
+  ];
+}
+
+export function shouldStartWithCustomModelInput(
+  provider: OpenWikiProvider,
+): boolean {
+  return getProviderModelOptions(provider).length === 0;
+}
+
+export function getSelectedModelId(
+  provider: OpenWikiProvider,
+  selectedIndex: number,
+  input: string,
+  isCustomInput: boolean,
+): string | null {
+  if (!isCustomInput) {
+    const selectedOption = getModelSelectionOptions(provider)[selectedIndex];
+
+    if (!selectedOption) {
+      return null;
+    }
+
+    return selectedOption.kind === "custom" ? "custom" : selectedOption.id;
+  }
+
+  const normalizedModelId = normalizeModelId(input);
+
+  return isValidModelId(normalizedModelId) ? normalizedModelId : null;
+}
+
+export function getModelSelectionIndex(
+  provider: OpenWikiProvider,
+  selectedModelId: string,
+): number {
+  const selectedIndex = getModelSelectionOptions(provider).findIndex(
+    (option) => option.kind === "preset" && option.id === selectedModelId,
+  );
+
+  return selectedIndex === -1 ? 0 : selectedIndex;
+}
