+++ src/setup/credentials.tsx
@@ -1,18 +1,9 @@
-import { existsSync } from "node:fs";
-import { spawn } from "node:child_process";
 import React, { useEffect, useMemo, useRef, useState } from "react";
-import { homedir } from "node:os";
 import path from "node:path";
 import { Box, Text, useInput, useStdin, useStdout } from "ink";
 import { configureAuthProvider } from "../auth/configure.js";
 import { runOAuthAuth } from "../auth/oauth.js";
 import {
-  AWS_ACCESS_KEY_ID_ENV_KEY,
-  AWS_BEARER_TOKEN_BEDROCK_ENV_KEY,
-  AWS_SECRET_ACCESS_KEY_ENV_KEY,
-  AWS_SESSION_TOKEN_ENV_KEY,
-  BEDROCK_AWS_ACCESS_KEY_ID_ENV_KEY,
-  BEDROCK_AWS_SECRET_ACCESS_KEY_ENV_KEY,
   DEFAULT_PROVIDER,
   DEFAULT_VERTEX_LOCATION,
   getDefaultModelId,
@@ -22,23 +13,12 @@ import {
   getProviderBaseUrlWarnings,
   getProviderLabel,
   getProviderLocationEnvKey,
-  getProviderModelOptions,
   getProviderProjectEnvKey,
   getProviderRegionEnvKey,
   getProviderRegionEnvKeys,
   getProviderSecretKeyEnvKey,
-  providerRequiresApiKey,
-  isValidModelId,
-  normalizeProvider,
-  normalizeModelId,
-  OPENAI_CHATGPT_EMAIL_ENV_KEY,
-  OPENAI_CHATGPT_PLAN_ENV_KEY,
-  OPENWIKI_GOOGLE_CLIENT_ID_ENV_KEY,
-  OPENWIKI_GOOGLE_CLIENT_SECRET_ENV_KEY,
   OPENWIKI_MODEL_ID_ENV_KEY,
   OPENWIKI_PROVIDER_ENV_KEY,
-  OPENWIKI_TAVILY_API_KEY_ENV_KEY,
-  OPENWIKI_X_CLIENT_ID_ENV_KEY,
   type OpenWikiProvider,
   providerRequiresBaseUrl,
   providerRequiresRegion,
@@ -54,23 +34,17 @@ import {
   type ChatGptLoginHandle,
   type CodexTokens,
   codexTokensToEnv,
-  formatChatGptAccount,
-  isChatGptTokenExpired,
   loginWithChatGPT,
-  readCodexTokensFromEnv,
 } from "../agent/openai-chatgpt-oauth.js";
-import type { AuthProviderId } from "../auth/types.js";
 import type { OpenWikiRunMode } from "../cli/commands.js";
 import {
   loadLangSmithSetup,
   nextLangSmithApiKeyEnv,
   saveLangSmithSetup,
 } from "../connectors/sources/langsmith/setup.js";
-import type { LangSmithRegion } from "../connectors/sources/langsmith/setup.js";
 import type { ConnectorId } from "../connectors/types.js";
 import {
   detectExternalCliCredential,
-  getExternalCliAuthAdapter,
   isExternalCliAvailable,
   runExternalCliLogin,
   type ExternalCliAuthState,
@@ -79,16 +53,12 @@ import { getConnectorConfigPath } from "../config/openwiki-home.js";
 import {
   getSavedEnvValue,
   getShellEnvValue,
-  openWikiEnvPath,
   saveOpenWikiEnv,
 } from "../config/env.js";
 import {
   createEmptyOnboardingConfig,
-  isOpenWikiOnboardingCompleteSync,
   isOnboardingComplete,
-  isRepositoryCodeOnboardingCompleteSync,
   readOpenWikiOnboardingConfig,
-  readRepositoryWikiInstructions,
   saveRepositoryWikiInstructions,
   saveOpenWikiOnboardingConfig,
   type OpenWikiOnboardingConfig,
@@ -99,6 +69,117 @@ import {
   installConnectorSchedule,
   validateCronExpression,
 } from "../scheduling/schedules.js";
+import {
+  credentialStep,
+  getConfigModeId,
+  getDefaultCodeRepoRootPath,
+  hasValidConfiguredProvider,
+  isBaseUrlConfigured,
+  isCodeMode,
+  isCredentialConfigured,
+  isRegionConfigured,
+  isScheduleStep,
+  isSecretKeyConfigured,
+  isSourceStep,
+  needsAwsCredentialRepair,
+  needsBaseUrlStep,
+  needsCredentialStep,
+  needsLangSmithStep,
+  needsRegionStep,
+  needsSecretKeyStep,
+  nextSetupStep,
+  type PromptStep,
+  resolveStepStatus,
+} from "./steps/machine.js";
+import {
+  getInitialStep,
+  getNextStepAfterApiKey,
+  getNextStepAfterBaseUrl,
+  getNextStepAfterGcpLocation,
+  getNextStepAfterProvider,
+  getNextStepAfterRegion,
+  getNextStepAfterSecretKey,
+} from "./steps/transitions.js";
+import {
+  ONBOARDING_TEMPLATES,
+  ensureRunModeConfig,
+  hydrateRunModeConfig,
+} from "./config/run-mode.js";
+import {
+  getModelSelectionIndex,
+  getModelSelectionOptions,
+  getModelSetupDetail,
+  getSelectedModelId,
+  shouldStartWithCustomModelInput,
+} from "./config/model-selection.js";
+import {
+  CODE_REPO_OPTIONS,
+  CRON_MODE_OPTIONS,
+  FINAL_OPTIONS,
+  getLangsmithRegionSelectionIndex,
+  getRunModeName,
+  getRunModeSelectionIndex,
+  getTemplateGoal,
+  LANGSMITH_REGION_OPTIONS,
+  POWER_MODE_OPTIONS,
+  RUN_MODE_OPTIONS,
+  SOURCE_CONTINUE_OPTIONS,
+} from "./config/options.js";
+import {
+  getInputDisplayWidth,
+  getProviderSelectionIndex,
+  moveSelectionIndex,
+} from "./display.js";
+import { copyToClipboard, openLoginUrl } from "./actions.js";
+import { handleCronEditorInput, sanitizeInputChunk } from "./inputs/cron.js";
+import {
+  getDefaultLocalGitRepoPath,
+  normalizeLocalPath,
+  sanitizeRepoId,
+  validateLocalDirectoryPath,
+} from "./inputs/local-path.js";
+import {
+  addSourceInstanceConfig,
+  createSourceInstanceId,
+  createSourceInstanceName,
+  getConnectedSourceCount,
+  getSourceDescriptionOptionCount,
+  getSourceOption,
+  getStaticSourceConfig,
+  getTemplateSourceOptions,
+  needsEnvValue,
+  type SourceSetupOption,
+} from "./sources.js";
+import {
+  getAwsCredentialRepairMessage,
+  getCredentialSetupDetail,
+  getWizardManagedEnvKeys,
+} from "./credential-detail.js";
+import { SetupHeader, SetupPanel, SetupStep } from "./components/layout.js";
+import { OAuthLoginPrompt } from "./components/oauth.js";
+import { Prompt } from "./components/prompt.js";
+import type {
+  LangsmithWorkspaceDraft,
+  PromptInputKey,
+  SourceSetupState,
+} from "./types.js";
+
+export {
+  getOAuthAuthorizationStatusText,
+  needsLangSmithStep,
+  nextSetupStep,
+  resolveStepStatus,
+} from "./steps/machine.js";
+export { needsCredentialSetup, orderedSetupSteps } from "./steps/machine.js";
+export { findNearestGitRepoRoot } from "./steps/machine.js";
+export {
+  getInitialStep,
+  getNextStepAfterProvider,
+} from "./steps/transitions.js";
+export {
+  ensureRunModeConfig,
+  hydrateRunModeConfig,
+} from "./config/run-mode.js";
 
 export type InitSetupResult = {
   mode: OpenWikiRunMode;
@@ -132,697 +213,6 @@ type InitSetupProps = {
    */
   walkAllSteps?: boolean;
 };
-
-type PromptStep =
-  | "api-key"
-  | "base-url"
-  | "code-repo-confirm"
-  | "code-repo-path"
-  | "external-cli-auth"
-  | "final"
-  | "gcp-location"
-  | "gcp-project"
-  | "langsmith"
-  | "model"
-  | "oauth-login"
-  | "provider"
-  | "region"
-  | "run-mode"
-  | "secret-key"
-  | "source-auth"
-  | "global-cron-custom"
-  | "global-cron-mode"
-  | "global-power-mode"
-  | "source-description"
-  | "source-description-custom"
-  | "source-langsmith-key"
-  | "source-langsmith-projects"
-  | "source-langsmith-region"
-  | "source-langsmith-workspaces"
-  | "source-menu"
-  | "source-path"
-  | "source-confirm-continue"
-  | "source-secret"
-  | "template"
-  | "wiki-goal";
-
-type SourceSetupOption = {
-  authProvider?: AuthProviderId;
-  displayName: string;
-  examples: string[];
-  id: ConnectorId;
-  instructions: string[];
-  secretInputs: SourceSecretInput[];
-};
-
-type SourceSecretInput = {
-  envKey: string;
-  label: string;
-  optional?: boolean;
-  secret?: boolean;
-};
-
-type SourceSetupState = {
-  authUrl?: string;
-  connectorConfig?: Record<string, unknown>;
-  copiedAuthUrlToClipboard?: boolean;
-  savedScheduleWarning?: string;
-  secretValues: Record<string, string>;
-};
-
-type PromptInputKey = {
-  backspace?: boolean;
-  ctrl?: boolean;
-  delete?: boolean;
-  downArrow?: boolean;
-  leftArrow?: boolean;
-  meta?: boolean;
-  return?: boolean;
-  rightArrow?: boolean;
-  tab?: boolean;
-  upArrow?: boolean;
-};
-
-type ModelSelectionOption =
-  | {
-      id: string;
-      kind: "preset";
-      label: string;
-    }
-  | {
-      kind: "custom";
-    };
-
-type OnboardingMode = {
-  description: string;
-  id: string;
-  name: string;
-  sourceIds: ConnectorId[];
-  suggestedSources: string[];
-  suggestedGoal: string;
-};
-
-const ONBOARDING_TEMPLATES = [
-  {
-    description:
-      "Maintain a structured project wiki from a local Git repository, with code-oriented pages for architecture, workflows, source maps, and operational guidance.",
-    id: "code",
-    name: "Code",
-    sourceIds: ["langsmith"],
-    suggestedSources: ["Local Git repository"],
-    suggestedGoal:
-      "A code wiki for this local repository. Prioritize a concise quickstart, architecture overview, source map, key workflows, domain concepts, operations/runbook notes, testing guidance, and integration points. Inspect git history to understand reasoning behind code changes and the progression of the repository. Keep pages grounded in the repository structure and recent code changes. Prefer practical navigation for engineers over generic summaries.",
-  },
-  {
-    description:
-      "A personal assistant wiki that builds memory from email, notes, social/research sources, and web search so you can ask about projects, priorities, people, and recurring context.",
-    id: "personal",
-    name: "Personal",
-    sourceIds: [
-      "git-repo",
-      "google",
-      "notion",
-      "web-search",
-      "hackernews",
-      "x",
-    ],
-    suggestedSources: [
-      "Gmail",
-      "Notion",
-      "Web Search (Tavily)",
-      "Hacker News",
-      "X/Twitter",
-    ],
-    suggestedGoal:
-      "Your personal brain. Track active projects, people, organizations, decisions, commitments, follow-ups, useful links, recurring themes, and fresh external signals. Organize the wiki so a personal assistant can answer what changed, what matters, what needs attention, and where supporting evidence came from. Be selective: summarize durable context and explicit action items, not every raw item.",
-  },
-] as const satisfies readonly OnboardingMode[];
-
-const RUN_MODE_OPTIONS = [
-  {
-    description:
-      "Build a local personal brain wiki in ~/.openwiki/wiki from configured sources.",
-    id: "personal",
-    name: "Personal",
-  },
-  {
-    description:
-      "Build repository documentation in ./openwiki for this codebase.",
-    id: "code",
-    name: "Code",
-  },
-] as const satisfies readonly {
-  description: string;
-  id: OpenWikiRunMode;
-  name: string;
-}[];
-
-const LANGSMITH_REGION_OPTIONS = [
-  {
-    description: "US workspaces. The default.",
-    host: "https://api.smith.langchain.com",
-    id: "us",
-    name: "US",
-  },
-  {
-    description: "EU workspaces.",
-    host: "https://eu.api.smith.langchain.com",
-    id: "eu",
-    name: "EU",
-  },
-] as const satisfies readonly {
-  description: string;
-  host: string;
-  id: LangSmithRegion;
-  name: string;
-}[];
-
-/**
- * One LangSmith workspace as the wizard edits it. `apiKey` holds a value entered
- * this session (empty = keep the committed key); it is written to ~/.openwiki/.env
- * under `apiKeyEnv` on completion, never committed.
- */
-interface LangsmithWorkspaceDraft {
-  apiKeyEnv: string;
-  region: LangSmithRegion;
-  apiKey: string;
-  projects: string[];
-}
-
-const SOURCE_OPTIONS = [
-  {
-    displayName: "Local Git repository",
-    examples: [
-      "Track architecture notes from this repo.",
-      "Summarize recent commits and changed files.",
-    ],
-    id: "git-repo",
-    instructions: [
-      "Choose the local repository directory OpenWiki should read.",
-      "The default is the current working directory, and you can replace it with another path.",
-      "You can add more repositories later in the connector config file.",
-    ],
-    secretInputs: [],
-  },
-  {
-    displayName: "LangSmith traces",
-    examples: ["support-bot-prod", "chat-agent"],
-    id: "langsmith",
-    instructions: [
-      "Document how your agent runs, grounded in its LangSmith traces.",
-      "List the projects to document; written to openwiki/.langsmith.json (committed).",
-    ],
-    // No secret input: the LangSmith key is captured by the earlier `langsmith`
-    // spine step (and provided as a CI secret), and used at pull time, not here.
-    secretInputs: [],
-  },
-  {
-    authProvider: "notion",
-    displayName: "Notion",
-    examples: [
-      "Ingest product specs, meeting notes, and research pages.",
-      "Prioritize pages related to Applied AI and customer feedback.",
-    ],
-    id: "notion",
-    instructions: [
-      "OpenWiki uses Notion's hosted MCP OAuth flow.",
-      "No client ID, client secret, or pasted Notion token is required.",
-      "Approve access in the browser window when it opens.",
-    ],
-    secretInputs: [],
-  },
-  {
-    authProvider: "gmail",
-    displayName: "Gmail",
-    examples: [
-      "Capture important project email threads from the last 24 hours.",
-      "Look for vendor updates, customer feedback, and action items.",
-    ],
-    id: "google",
-    instructions: [
-      "Create OAuth credentials in Google Cloud for a desktop or web app.",
-      "Enable the Gmail API for the Google Cloud project.",
-      "Add http://127.0.0.1:53682/callback as an authorized redirect URI.",
-      "Paste the client ID and client secret below.",
-    ],
-    secretInputs: [
-      {
-        envKey: OPENWIKI_GOOGLE_CLIENT_ID_ENV_KEY,
-        label: "Google OAuth client ID",
-      },
-      {
-        envKey: OPENWIKI_GOOGLE_CLIENT_SECRET_ENV_KEY,
-        label: "Google OAuth client secret",
-        secret: true,
-      },
-    ],
-  },
-  {
-    displayName: "Web Search (Tavily)",
-    examples: [
-      "Track a company, product category, or technical topic.",
-      "Find launch posts, docs, pricing pages, and recent articles.",
-    ],
-    id: "web-search",
-    instructions: [
-      "Create a Tavily account and API key.",
-      "Paste the Tavily API key below.",
-      "Describe the topics, companies, or pages OpenWiki should search for on the next screen.",
-    ],
-    secretInputs: [
-      {
-        envKey: OPENWIKI_TAVILY_API_KEY_ENV_KEY,
-        label: "Tavily API key",
-        secret: true,
-      },
-    ],
-  },
-  {
-    displayName: "Hacker News",
-    examples: [
-      "Monitor threads about AI agents, evals, infrastructure, and startups.",
-      "Capture notable discussions and links related to my research topics.",
-    ],
-    id: "hackernews",
-    instructions: [
-      "No account setup is required for Hacker News.",
-      "OpenWiki uses public Hacker News feed and search APIs.",
-      "Describe the topics, keywords, users, or story types OpenWiki should watch on the next screen.",
-    ],
-    secretInputs: [],
-  },
-  {
-    authProvider: "x",
-    displayName: "X / Twitter",
-    examples: [
-      "Track my home timeline, bookmarks, and key lists.",
-      "Summarize tweets from AI researchers and product announcements.",
-    ],
-    id: "x",
-    instructions: [
-      "Create an X OAuth 2.0 app.",
-      "Use a native app or public client when possible.",
-      "Add http://127.0.0.1:53682/callback as a callback URI.",
-      "Paste the OAuth client ID below.",
-    ],
-    secretInputs: [
-      {
-        envKey: OPENWIKI_X_CLIENT_ID_ENV_KEY,
-        label: "X OAuth client ID",
-      },
-    ],
-  },
-] as const satisfies readonly SourceSetupOption[];
-
-const CRON_MODE_OPTIONS = [
-  "Use suggested schedule",
-  "Enter custom cron",
-] as const;
-const POWER_MODE_OPTIONS = [
-  "Set up Mac wake/sleep window",
-  "Skip power setup",
-] as const;
-const CRON_FIELD_LABELS = ["minute", "hour", "day", "month", "weekday"];
-const SOURCE_CONTINUE_OPTIONS = [
-  "Go back to connections",
-  "Continue without all sources",
-] as const;
-const FINAL_OPTIONS = ["Run ingestion now", "Run later"] as const;
-const CODE_REPO_OPTIONS = ["Confirm and continue", "Edit path"] as const;
-
-export function needsCredentialSetup(
-  modelIdOverride: string | null = null,
-  mode: OpenWikiRunMode = "personal",
-): boolean {
-  const provider = resolveConfiguredProvider();
-
-  const needsCredentials =
-    !hasValidConfiguredProvider() ||
-    needsAwsCredentialRepair(provider) ||
-    needsCredentialStep(provider) ||
-    needsSecretKeyStep(provider) ||
-    needsBaseUrlStep(provider) ||
-    needsRegionStep(provider) ||
-    (modelIdOverride === null &&
-      process.env[OPENWIKI_MODEL_ID_ENV_KEY] === undefined) ||
-    needsLangSmithStep();
-
-  if (needsCredentials) {
-    return true;
-  }
-
-  return mode === "code"
-    ? !isRepositoryCodeOnboardingCompleteSync(getDefaultCodeRepoRootPath())
-    : !isOpenWikiOnboardingCompleteSync();
-}
-
-function needsAwsCredentialRepair(provider: OpenWikiProvider): boolean {
-  return (
-    providerUsesAwsSdkCredentials(provider) &&
-    getMissingProviderEnvKey(provider) !== null
-  );
-}
-
-function getAwsCredentialRepairMessage(
-  provider: OpenWikiProvider,
-): string | null {
-  if (!providerUsesAwsSdkCredentials(provider)) {
-    return null;
-  }
-
-  const missingEnvKey = getMissingProviderEnvKey(provider);
-
-  if (!missingEnvKey) {
-    return null;
-  }
-
-  const pair =
-    missingEnvKey === BEDROCK_AWS_ACCESS_KEY_ID_ENV_KEY ||
-    missingEnvKey === BEDROCK_AWS_SECRET_ACCESS_KEY_ENV_KEY
-      ? `${BEDROCK_AWS_ACCESS_KEY_ID_ENV_KEY} and ${BEDROCK_AWS_SECRET_ACCESS_KEY_ENV_KEY}`
-      : `${AWS_ACCESS_KEY_ID_ENV_KEY} and ${AWS_SECRET_ACCESS_KEY_ENV_KEY}`;
-
-  return `${missingEnvKey} is missing or blank. Set both ${pair}, or unset both in your shell and ${openWikiEnvPath}, then restart OpenWiki.`;
-}
-
-/**
- * Whether the provider still needs its primary credential collected. For
- * `oauth` providers this is a valid, non-expired stored token; for API-key
- * providers it is a pasted key; for keyless providers (gemini-enterprise) it is
- * the required GCP project id.
- */
-function needsCredentialStep(provider: OpenWikiProvider): boolean {
-  if (providerUsesOAuth(provider)) {
-    return !hasValidStoredToken();
-  }
-
-  return (
-    getMissingProviderEnvKey(provider) !== null &&
-    credentialStep(provider) !== null
-  );
-}
-
-/** The step that collects the provider's primary credential. */
-function credentialStep(provider: OpenWikiProvider): PromptStep | null {
-  if (providerUsesOAuth(provider)) {
-    return "oauth-login";
-  }
-
-  if (providerUsesAwsSdkCredentials(provider)) {
-    return null;
-  }
-
-  if (providerUsesExternalCliAuth(provider)) {
-    return "external-cli-auth";
-  }
-
-  if (providerRequiresApiKey(provider)) {
-    return "api-key";
-  }
-
-  return getProviderProjectEnvKey(provider) ? "gcp-project" : null;
-}
-
-/**
- * Every managed env key the wizard lets you set for a provider, in checklist
- * order: the provider selection, its credential keys, the model, and the
- * LangSmith tracing key. Used to detect which of them a shell export is
- * currently shadowing (a shell var wins at runtime and would silently override
- * the choice made here). Returns key names only, never values.
- */
-function getWizardManagedEnvKeys(provider: OpenWikiProvider): string[] {
-  return [
-    OPENWIKI_PROVIDER_ENV_KEY,
-    getProviderApiKeyEnvKey(provider),
-    getProviderSecretKeyEnvKey(provider),
-    getProviderProjectEnvKey(provider),
-    getProviderLocationEnvKey(provider),
-    getProviderBaseUrlEnvKey(provider),
-    getProviderRegionEnvKey(provider),
-    OPENWIKI_MODEL_ID_ENV_KEY,
-    "LANGSMITH_API_KEY",
-  ].filter((key): key is string => key !== undefined);
-}
-
-/**
- * The setup steps that apply to a provider and run mode, in the order the wizard
- * walks them. Unlike the skip-based waterfall in {@link getInitialStep}, this
- * includes steps already satisfied by the environment, so navigation can reach
- * and re-edit an auto-skipped step. The provider's primary credential step
- * ({@link credentialStep}) is emitted once; for keyless providers that step is
- * the GCP project, so it is not appended again below.
- */
-export function orderedSetupSteps(
-  provider: OpenWikiProvider,
-  mode: OpenWikiRunMode,
-  allowModeSelection: boolean,
-): PromptStep[] {
-  const steps: PromptStep[] = [];
-
-  if (allowModeSelection) {
-    steps.push("run-mode");
-  }
-
-  steps.push("provider");
-
-  const primary = credentialStep(provider);
-  if (primary) {
-    steps.push(primary);
-  }
-
-  if (providerRequiresSecretKey(provider)) {
-    steps.push("secret-key");
-  }
-  if (getProviderProjectEnvKey(provider) && primary !== "gcp-project") {
-    steps.push("gcp-project");
-  }
-  if (
-    getProviderProjectEnvKey(provider) &&
-    getProviderLocationEnvKey(provider)
-  ) {
-    steps.push("gcp-location");
-  }
-  if (providerRequiresBaseUrl(provider)) {
-    steps.push("base-url");
-  }
-  if (providerRequiresRegion(provider)) {
-    steps.push("region");
-  }
-
-  steps.push("model");
-  steps.push("langsmith");
-
-  // Personal mode's template is fixed by the run mode, so it skips the
-  // Code/Personal chooser and walks straight into the wiki brief. Only code
-  // mode needs a spine step after langsmith (repo confirmation).
-  if (mode === "code") {
-    steps.push("code-repo-confirm");
-  }
-
-  return steps;
-}
-
-/**
- * The step after `step` in the applicable spine, or null when `step` is the last
- * spine step or outside it. Drives forward navigation: Enter advances to the
- * next applicable step in order rather than skipping ones already satisfied by
- * the environment, so setup reads as a sequential walk.
- */
-export function nextSetupStep(
-  step: PromptStep | null,
-  provider: OpenWikiProvider,
-  mode: OpenWikiRunMode,
-  allowModeSelection: boolean,
-): PromptStep | null {
-  if (step === null) {
-    return null;
-  }
-  const spine = orderedSetupSteps(provider, mode, allowModeSelection);
-  const index = spine.indexOf(step);
-  return index >= 0 && index + 1 < spine.length ? spine[index + 1] : null;
-}
-
-function hasValidStoredToken(env: NodeJS.ProcessEnv = process.env): boolean {
-  const tokens = readCodexTokensFromEnv(env);
-
-  return tokens !== null && !isChatGptTokenExpired(tokens.expiresAtMs);
-}
-
-function needsGcpProjectStep(provider: OpenWikiProvider): boolean {
-  const projectEnvKey = getProviderProjectEnvKey(provider);
-
-  return projectEnvKey ? !process.env[projectEnvKey] : false;
-}
-
-function needsBaseUrlStep(provider: OpenWikiProvider): boolean {
-  if (!providerRequiresBaseUrl(provider)) {
-    return false;
-  }
-
-  return !isBaseUrlConfigured(provider);
-}
-
-function isBaseUrlConfigured(provider: OpenWikiProvider): boolean {
-  const baseUrlEnvKey = getProviderBaseUrlEnvKey(provider);
-
-  return baseUrlEnvKey ? Boolean(process.env[baseUrlEnvKey]) : false;
-}
-
-function needsSecretKeyStep(provider: OpenWikiProvider): boolean {
-  if (!providerRequiresSecretKey(provider)) {
-    return false;
-  }
-
-  return !isSecretKeyConfigured(provider);
-}
-
-function isSecretKeyConfigured(provider: OpenWikiProvider): boolean {
-  const secretKeyEnvKey = getProviderSecretKeyEnvKey(provider);
-
-  return secretKeyEnvKey ? Boolean(process.env[secretKeyEnvKey]) : false;
-}
-
-function needsRegionStep(provider: OpenWikiProvider): boolean {
-  if (!providerRequiresRegion(provider)) {
-    return false;
-  }
-
-  return !isRegionConfigured(provider);
-}
-
-/**
- * Whether the optional LangSmith tracing step still needs to be shown.
- *
- * The step is optional, so "answered" must include skipping it. Skipping does
- * not persist `LANGSMITH_API_KEY` — `saveOpenWikiEnv` strips empty values, so
- * the key is simply absent afterwards. What the step always records instead is
- * `LANGCHAIN_TRACING_V2` (`"false"` on skip, `"true"` when a key is entered),
- * which survives because it is non-empty. So the step is unanswered only when
- * neither a key is present (e.g. from a shell export) nor a tracing decision
- * has been recorded.
- */
-export function needsLangSmithStep(
-  env: NodeJS.ProcessEnv = process.env,
-): boolean {
-  return !env.LANGSMITH_API_KEY && env.LANGCHAIN_TRACING_V2 === undefined;
-}
-
-function isRegionConfigured(provider: OpenWikiProvider): boolean {
-  return resolveProviderRegion(provider) !== undefined;
-}
-
-function isCredentialConfigured(provider: OpenWikiProvider): boolean {
-  return providerUsesOAuth(provider)
-    ? hasValidStoredToken()
-    : getMissingProviderEnvKey(provider) === null;
-}
-
-function getCredentialSetupDetail(
-  provider: OpenWikiProvider,
-  tokens: CodexTokens | null = null,
-): string {
-  if (providerUsesOAuth(provider)) {
-    if (!isCredentialConfigured(provider) && !tokens) {
-      return "sign in with your ChatGPT account";
-    }
-
-    const account = formatChatGptAccount(
-      tokens?.email ?? process.env[OPENAI_CHATGPT_EMAIL_ENV_KEY] ?? null,
-      tokens?.planType ?? process.env[OPENAI_CHATGPT_PLAN_ENV_KEY] ?? null,
-    );
-
-    return account ? `signed in as ${account}` : "signed in with ChatGPT";
-  }
-
-  if (providerUsesAwsSdkCredentials(provider)) {
-    if (process.env[AWS_BEARER_TOKEN_BEDROCK_ENV_KEY]?.trim()) {
-      return "Bedrock bearer token (takes precedence)";
-    }
-
-    const missingEnvKey = getMissingProviderEnvKey(provider);
-
-    if (missingEnvKey) {
-      if (
-        missingEnvKey === BEDROCK_AWS_ACCESS_KEY_ID_ENV_KEY ||
-        missingEnvKey === BEDROCK_AWS_SECRET_ACCESS_KEY_ENV_KEY
-      ) {
-        return "incomplete legacy Bedrock keys; set both or clear both";
-      }
-
-      if (
-        missingEnvKey === AWS_ACCESS_KEY_ID_ENV_KEY ||
-        missingEnvKey === AWS_SECRET_ACCESS_KEY_ENV_KEY
-      ) {
-        return "incomplete standard AWS credentials; set the full set or unset it";
-      }
-
-      return `incomplete AWS credential configuration (${missingEnvKey})`;
-    }
-
-    const legacyApiKey = getProviderApiKeyEnvKey(provider);
-    const legacySecretKey = getProviderSecretKeyEnvKey(provider);
-    const usesLegacyKeys = Boolean(
-      legacyApiKey &&
-      legacySecretKey &&
-      process.env[legacyApiKey]?.trim() &&
-      process.env[legacySecretKey]?.trim(),
-    );
-
-    const ignoresOrphanSessionToken = Boolean(
-      process.env[AWS_SESSION_TOKEN_ENV_KEY]?.trim() &&
-      !process.env[AWS_ACCESS_KEY_ID_ENV_KEY]?.trim() &&
-      !process.env[AWS_SECRET_ACCESS_KEY_ENV_KEY]?.trim(),
-    );
-
-    return usesLegacyKeys
-      ? "legacy Bedrock keys (take precedence)"
-      : ignoresOrphanSessionToken
-        ? "AWS SDK default credential chain (orphan AWS_SESSION_TOKEN ignored)"
-        : "AWS SDK default credential chain";
-  }
-
-  const apiKeyEnvKey = getProviderApiKeyEnvKey(provider);
-
-  return isCredentialConfigured(provider)
-    ? "available from environment"
-    : apiKeyEnvKey
-      ? `save ${apiKeyEnvKey} to ${openWikiEnvPath}`
-      : "configure Google Cloud credentials";
-}
-
-/**
- * Copies text to the terminal's clipboard using the OSC 52 escape sequence.
- * This targets the user's local terminal emulator even when OpenWiki runs over
- * SSH, unlike shelling out to a host clipboard utility.
- */
-function copyToClipboard(text: string): void {
-  const encoded = Buffer.from(text, "utf8").toString("base64");
-
-  process.stdout.write(`\u001b]52;c;${encoded}\u0007`);
-}
-
-function openLoginUrl(url: string): void {
-  try {
-    const child =
-      process.platform === "win32"
-        ? spawn("cmd", ["/c", "start", '""', `"${url}"`], {
-            detached: true,
-            stdio: "ignore",
-            windowsVerbatimArguments: true,
-          })
-        : spawn(process.platform === "darwin" ? "open" : "xdg-open", [url], {
-            detached: true,
-            stdio: "ignore",
-          });
-
-    child.on("error", () => {
-      // The URL is also rendered for manual use on headless/SSH machines.
-    });
-    child.unref();
-  } catch {
-    // Ignore spawn failures; the URL is still rendered for manual use.
-  }
-}
 
 export function InitSetup({
   allowModeSelection = false,
@@ -3605,2115 +2995,6 @@ export function InitSetup({
       ) : null}
     </Box>
   );
-}
-
-function Prompt({
-  codeRepoPathInput,
-  codeRepoRoot,
-  codeRepoSelectionIndex,
-  externalCliAuth,
-  cronFieldSelectionIndex,
-  cronModeSelectionIndex,
-  finalSelectionIndex,
-  input,
-  inputDisplayWidth,
-  isCustomModelInput,
-  langsmithDraft,
-  langsmithRegionSelectionIndex,
-  langsmithWorkspaceSelectionIndex,
-  langsmithWorkspaces,
-  modelSelectionIndex,
-  onboardingConfig,
-  powerModeSelectionIndex,
-  provider,
-  providerSelectionIndex,
-  runModeSelectionIndex,
-  secretInputIndex,
-  selectedMode,
-  selectedSource,
-  sourceOptions,
-  sourceContinueSelectionIndex,
-  sourceDescriptionSelectionIndex,
-  sourceSelectionIndex,
-  sourceState,
-  step,
-  suggestedCronDescription,
-  suggestedCronExpression,
-  templateSelectionIndex,
-}: {
-  codeRepoPathInput: string;
-  codeRepoRoot: string;
-  codeRepoSelectionIndex: number;
-  externalCliAuth: ExternalCliAuthState;
-  cronFieldSelectionIndex: number;
-  cronModeSelectionIndex: number;
-  finalSelectionIndex: number;
-  input: string;
-  inputDisplayWidth: number;
-  isCustomModelInput: boolean;
-  langsmithDraft: LangsmithWorkspaceDraft | null;
-  langsmithRegionSelectionIndex: number;
-  langsmithWorkspaceSelectionIndex: number;
-  langsmithWorkspaces: LangsmithWorkspaceDraft[];
-  modelSelectionIndex: number;
-  onboardingConfig: OpenWikiOnboardingConfig;
-  powerModeSelectionIndex: number;
-  provider: OpenWikiProvider;
-  providerSelectionIndex: number;
-  runModeSelectionIndex: number;
-  secretInputIndex: number;
-  selectedMode: OpenWikiRunMode;
-  selectedSource: SourceSetupOption;
-  sourceOptions: readonly SourceSetupOption[];
-  sourceContinueSelectionIndex: number;
-  sourceDescriptionSelectionIndex: number;
-  sourceSelectionIndex: number;
-  sourceState: SourceSetupState;
-  step: PromptStep;
-  suggestedCronDescription: string;
-  suggestedCronExpression: string;
-  templateSelectionIndex: number;
-}) {
-  if (step === "run-mode") {
-    const selectedMode =
-      RUN_MODE_OPTIONS[runModeSelectionIndex] ?? RUN_MODE_OPTIONS[0];
-
-    return (
-      <Box flexDirection="column">
-        <Text>Choose what OpenWiki should initialize.</Text>
-        {RUN_MODE_OPTIONS.map((option, index) => (
-          <Text key={option.id}>
-            <SelectionMarker isSelected={index === runModeSelectionIndex} />{" "}
-            {option.name} <Text color="gray">({option.id})</Text>
-          </Text>
-        ))}
-        <Box flexDirection="column" marginTop={1}>
-          <Text bold>{selectedMode.name}</Text>
-          <Text color="gray">{selectedMode.description}</Text>
-        </Box>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "provider") {
-    return (
-      <Box flexDirection="column">
-        <Text>Choose a model provider.</Text>
-        {SELECTABLE_OPENWIKI_PROVIDERS.map((providerOption, index) => (
-          <Text key={providerOption}>
-            <SelectionMarker isSelected={index === providerSelectionIndex} />{" "}
-            {getProviderLabel(providerOption)}
-            <Text color="gray"> ({providerOption})</Text>
-            {providerOption === DEFAULT_PROVIDER ? (
-              <Text color="gray"> default</Text>
-            ) : null}
-          </Text>
-        ))}
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "api-key") {
-    return (
-      <Box flexDirection="column">
-        <Text>Paste your {getApiKeyFieldLabel(provider)}.</Text>
-        <BorderedInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          prefix={`${getProviderApiKeyEnvKey(provider)}=`}
-          secret
-          value={input}
-        />
-        <Text color="gray">Press Enter to save it.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "external-cli-auth") {
-    return (
-      <ExternalCliAuthPrompt
-        authState={externalCliAuth}
-        input={input}
-        provider={provider}
-      />
-    );
-  }
-
-  if (step === "secret-key") {
-    return (
-      <Box flexDirection="column">
-        <Text>Paste your {getProviderLabel(provider)} secret access key.</Text>
-        <BorderedInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          prefix={`${getProviderSecretKeyEnvKey(provider)}=`}
-          secret
-          value={input}
-        />
-        <Text color="gray">Press Enter to save it.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "gcp-project") {
-    return (
-      <Box flexDirection="column">
-        <Text>Enter the Google Cloud project ID with Vertex AI access.</Text>
-        <Text>
-          <Text color="gray">$</Text> {getProviderProjectEnvKey(provider)}={" "}
-          <Text color="yellow">{input}</Text>
-        </Text>
-        <Text color="gray">
-          OpenWiki authenticates with Google Application Default Credentials
-          (run: gcloud auth application-default login). Press Enter to save it.
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "gcp-location") {
-    return (
-      <Box flexDirection="column">
-        <Text>
-          Enter a Vertex AI location, or press Enter to use{" "}
-          {DEFAULT_VERTEX_LOCATION}.
-        </Text>
-        <Text>
-          <Text color="gray">$</Text> {getProviderLocationEnvKey(provider)}={" "}
-          <Text color="yellow">{input}</Text>
-        </Text>
-        <Text color="gray">
-          For example global, europe-west1, or us-east5. Press Enter to
-          continue.
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "base-url") {
-    return (
-      <Box flexDirection="column">
-        <Text>Enter the {getProviderLabel(provider)} base URL.</Text>
-        <Text>
-          <Text color="gray">$</Text> {getProviderBaseUrlEnvKey(provider)}={" "}
-          <Text color="yellow">{input}</Text>
-        </Text>
-        <Text color="gray">
-          For example an OpenAI-compatible gateway endpoint (such as a LiteLLM
-          gateway). Press Enter to save it.
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "region") {
-    const resolvedRegion = resolveProviderRegion(provider);
-    const credentialRepairMessage = getAwsCredentialRepairMessage(provider);
-
-    return (
-      <Box flexDirection="column">
-        {credentialRepairMessage ? (
-          <Text color="yellow">⚠ {credentialRepairMessage}</Text>
-        ) : null}
-        <Text>
-          Enter the {getProviderLabel(provider)} region
-          {resolvedRegion ? `, or press Enter to keep ${resolvedRegion}` : ""}.
-        </Text>
-        <Text>
-          <Text color="gray">$</Text> {getProviderRegionEnvKey(provider)}={" "}
-          <Text color="yellow">{input}</Text>
-        </Text>
-        <Text color="gray">
-          Uses {getProviderRegionEnvKeys(provider).join(", ")}. For example
-          us-east-1.
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "model") {
-    if (isCustomModelInput) {
-      return (
-        <Box flexDirection="column">
-          <Text>Paste a custom model ID.</Text>
-          <BorderedInput
-            maxDisplayWidth={inputDisplayWidth}
-            marginTop={1}
-            prefix={`${OPENWIKI_MODEL_ID_ENV_KEY}=`}
-            value={input}
-          />
-          <Text color="gray">Press Enter to save it.</Text>
-        </Box>
-      );
-    }
-
-    return (
-      <Box flexDirection="column">
-        <Text>
-          Choose {getProviderArticle(provider)} {getProviderLabel(provider)}{" "}
-          model.
-        </Text>
-        {getModelSelectionOptions(provider).map((option, index) => {
-          if (option.kind === "custom") {
-            return (
-              <Text key="custom">
-                <SelectionMarker isSelected={index === modelSelectionIndex} />{" "}
-                Custom model ID
-              </Text>
-            );
-          }
-
-          return (
-            <Text key={option.id}>
-              <SelectionMarker isSelected={index === modelSelectionIndex} />{" "}
-              {option.label} <Text color="gray">{option.id}</Text>
-              {option.id === getDefaultModelId(provider) ? (
-                <Text color="gray"> default</Text>
-              ) : null}
-            </Text>
-          );
-        })}
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "langsmith") {
-    return (
-      <Box flexDirection="column">
-        <Text>Optional: paste a LangSmith API key for tracing.</Text>
-        <BorderedInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          prefix="LANGSMITH_API_KEY optional="
-          secret
-          value={input}
-        />
-        <Text color="gray">Press Enter with an empty value to skip.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "template") {
-    const selectedTemplate =
-      ONBOARDING_TEMPLATES[templateSelectionIndex] ?? ONBOARDING_TEMPLATES[0];
-
-    return (
-      <Box flexDirection="column">
-        <Text>Choose how OpenWiki should run.</Text>
-        {ONBOARDING_TEMPLATES.map((template, index) => (
-          <Text key={template.id}>
-            <SelectionMarker isSelected={index === templateSelectionIndex} />{" "}
-            {template.name}
-          </Text>
-        ))}
-        <Box flexDirection="column" marginTop={1}>
-          <Text bold>{selectedTemplate.name}</Text>
-          <Text color="gray">{selectedTemplate.description}</Text>
-          {selectedTemplate.suggestedSources.length > 0 ? (
-            <Text color="gray">
-              Suggested sources: {selectedTemplate.suggestedSources.join(", ")}
-            </Text>
-          ) : (
-            <Text color="gray">Start from a blank wiki brief.</Text>
-          )}
-        </Box>
-        <Text color="gray">
-          Press Enter, then edit the brief on the next step.
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "wiki-goal") {
-    return (
-      <Box flexDirection="column">
-        <Text>Customize what this wiki should understand.</Text>
-        {getConfigModeName(onboardingConfig) ? (
-          <Text color="gray">Mode: {getConfigModeName(onboardingConfig)}</Text>
-        ) : null}
-        <Text color="gray">
-          Edit the brief below. Keep what is useful, delete what is not.
-        </Text>
-        <Box flexDirection="column" marginTop={1}>
-          <Text bold>Edit wiki brief</Text>
-          <BorderedMultilineInput
-            maxDisplayWidth={inputDisplayWidth}
-            value={input}
-          />
-        </Box>
-        <Text color="gray">Press Enter to continue.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "code-repo-confirm") {
-    return (
-      <Box flexDirection="column">
-        <Text>Use this repository?</Text>
-        <Box marginTop={1}>
-          <Text color="cyan">{codeRepoRoot}</Text>
-        </Box>
-        <Text color="gray">
-          OpenWiki will run in this directory and write the initial openwiki/
-          folder there.
-        </Text>
-        <Box flexDirection="column" marginTop={1}>
-          {CODE_REPO_OPTIONS.map((option, index) => (
-            <Text key={option}>
-              <SelectionMarker isSelected={index === codeRepoSelectionIndex} />{" "}
-              {option}
-            </Text>
-          ))}
-        </Box>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "code-repo-path") {
-    return (
-      <Box flexDirection="column">
-        <Text>Choose the repository directory.</Text>
-        <Text color="gray">
-          Enter an existing directory. OpenWiki will write openwiki/ there.
-        </Text>
-        <BorderedInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          prefix="path="
-          value={codeRepoPathInput}
-        />
-        <Text color="gray">Press Enter to confirm this path.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-menu") {
-    // LangSmith workspaces live in state until setup completes (not onboarding
-    // source instances), so count them here for the menu's configured display.
-    const langsmithWorkspaceCount = sourceOptions.some(
-      (source) => source.id === "langsmith",
-    )
-      ? langsmithWorkspaces.length
-      : 0;
-    const configuredCount =
-      getConnectedSourceCount(onboardingConfig, sourceOptions) +
-      langsmithWorkspaceCount;
-
-    return (
-      <Box flexDirection="column">
-        <Text>Configure sources for this mode.</Text>
-        {sourceOptions.map((source, index) => {
-          const isLangsmith = source.id === "langsmith";
-          const sourceInstances = getSourceInstances(
-            onboardingConfig,
-            source.id,
-          );
-          const count = isLangsmith
-            ? langsmithWorkspaces.length
-            : sourceInstances.length;
-          return (
-            <Box flexDirection="column" key={source.id}>
-              <Text>
-                <SelectionMarker isSelected={index === sourceSelectionIndex} />{" "}
-                {getSourceMenuLabel(source, count)}{" "}
-                <SourceConnectionStatus
-                  count={count}
-                  isConfigured={count > 0}
-                />
-              </Text>
-              {isLangsmith
-                ? langsmithWorkspaces.map((workspace) => (
-                    <Text color="gray" key={workspace.apiKeyEnv}>
-                      {"  "}- {getLangsmithRegionLabel(workspace.region)}:{" "}
-                      {workspace.projects.join(", ")}
-                    </Text>
-                  ))
-                : sourceInstances.map((sourceInstance) => (
-                    <Text color="gray" key={sourceInstance.id}>
-                      {"  "}- {sourceInstance.name ?? sourceInstance.id}{" "}
-                      <Text color="gray">({sourceInstance.id})</Text>
-                    </Text>
-                  ))}
-            </Box>
-          );
-        })}
-        <Box flexDirection="column" marginTop={1}>
-          <Text color="gray">Next</Text>
-          <Text>
-            <SelectionMarker
-              isSelected={sourceSelectionIndex === sourceOptions.length}
-            />{" "}
-            Continue{" "}
-            {configuredCount === 0 ? (
-              <Text color="gray">(no sources configured)</Text>
-            ) : null}
-          </Text>
-        </Box>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-path") {
-    return (
-      <Box flexDirection="column">
-        <Text>Choose the local Git repository directory.</Text>
-        <Text color="gray">
-          Default is the directory where you started OpenWiki. Edit it to use a
-          different checkout.
-        </Text>
-        <BorderedInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          prefix="path="
-          value={input}
-        />
-        <Text color="gray">Press Enter to save this source.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-secret") {
-    const secretInput = selectedSource.secretInputs[secretInputIndex];
-    return (
-      <Box flexDirection="column">
-        <Text>{selectedSource.displayName} setup</Text>
-        {selectedSource.instructions.map((instruction, index) => (
-          <Text key={instruction}>
-            {index + 1}. {instruction}
-          </Text>
-        ))}
-        {secretInput ? (
-          <Box flexDirection="column" marginTop={1}>
-            <Text bold>Enter credential</Text>
-            <BorderedInput
-              maxDisplayWidth={inputDisplayWidth}
-              prefix={`${secretInput.envKey}${
-                secretInput.optional ? " optional" : ""
-              }=`}
-              secret
-              value={input}
-            />
-            <Text color="gray">
-              {secretInput.optional
-                ? "Press Enter with an empty value to skip."
-                : "Press Enter to save this value."}
-            </Text>
-          </Box>
-        ) : null}
-      </Box>
-    );
-  }
-
-  if (step === "source-auth") {
-    return (
-      <Box flexDirection="column">
-        <Text>{selectedSource.displayName} authorization</Text>
-        {sourceState.authUrl ? (
-          <OAuthAuthorizationLink
-            authProvider={selectedSource.authProvider}
-            copiedToClipboard={Boolean(sourceState.copiedAuthUrlToClipboard)}
-            url={sourceState.authUrl}
-          />
-        ) : (
-          <Text color="gray">
-            Press Enter to open the authorization URL and wait for the callback.
-          </Text>
-        )}
-      </Box>
-    );
-  }
-
-  if (step === "source-description") {
-    return (
-      <Box flexDirection="column">
-        <Text>{getSourceDescriptionPrompt(selectedSource)}</Text>
-        <Text color="gray">
-          Choose an example description, or write your own.
-        </Text>
-        {selectedSource.examples.map((example, index) => (
-          <Text key={example}>
-            <SelectionMarker
-              isSelected={index === sourceDescriptionSelectionIndex}
-            />{" "}
-            {example}
-          </Text>
-        ))}
-        <Text>
-          <SelectionMarker
-            isSelected={
-              sourceDescriptionSelectionIndex >= selectedSource.examples.length
-            }
-          />{" "}
-          Custom description
-        </Text>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-description-custom") {
-    return (
-      <Box flexDirection="column">
-        <Text>{getSourceDescriptionPrompt(selectedSource)}</Text>
-        <Text color="gray">
-          Type what OpenWiki should focus on for this source.
-        </Text>
-        <BorderedMultilineInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          value={input}
-        />
-        <Text color="gray">Optional. Press Enter to continue.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-langsmith-workspaces") {
-    const workspaceCount = langsmithWorkspaces.length;
-    return (
-      <Box flexDirection="column">
-        <Text>LangSmith workspaces to document.</Text>
-        <Text color="gray">
-          A LangSmith key is region-bound, so each workspace has its own region
-          and key. Select one to edit (clear its projects to remove it).
-        </Text>
-        {langsmithWorkspaces.map((workspace, index) => (
-          <Text key={workspace.apiKeyEnv}>
-            <SelectionMarker
-              isSelected={index === langsmithWorkspaceSelectionIndex}
-            />{" "}
-            {getLangsmithRegionLabel(workspace.region)}:{" "}
-            {workspace.projects.join(", ")}
-          </Text>
-        ))}
-        <Box flexDirection="column" marginTop={1}>
-          <Text>
-            <SelectionMarker
-              isSelected={langsmithWorkspaceSelectionIndex === workspaceCount}
-            />{" "}
-            Add a workspace
-          </Text>
-          <Text>
-            <SelectionMarker
-              isSelected={
-                langsmithWorkspaceSelectionIndex === workspaceCount + 1
-              }
-            />{" "}
-            Done
-          </Text>
-        </Box>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-langsmith-key") {
-    const apiKeyEnv = langsmithDraft?.apiKeyEnv ?? "OPENWIKI_LANGSMITH_API_KEY";
-    return (
-      <Box flexDirection="column">
-        <Text>LangSmith API key for this workspace.</Text>
-        <Text color="gray">
-          The connector&apos;s own read key (not your app&apos;s tracing key).
-          Saved to ~/.openwiki/.env as {apiKeyEnv}, never committed.
-        </Text>
-        <BorderedInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          prefix={`${apiKeyEnv}=`}
-          secret
-          value={input}
-        />
-        <Text color="gray">
-          Press Enter to confirm (empty keeps the saved key).
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-langsmith-projects") {
-    return (
-      <Box flexDirection="column">
-        <Text>Which projects should this wiki document in this workspace?</Text>
-        <Text color="gray">
-          Comma-separated project names (as in LANGCHAIN_PROJECT). Written to
-          openwiki/.langsmith.json.
-        </Text>
-        <BorderedMultilineInput
-          maxDisplayWidth={inputDisplayWidth}
-          marginTop={1}
-          value={input}
-        />
-        <Text color="gray">Press Enter to confirm.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-langsmith-region") {
-    const selectedRegion =
-      LANGSMITH_REGION_OPTIONS[langsmithRegionSelectionIndex] ??
-      LANGSMITH_REGION_OPTIONS[0];
-
-    return (
-      <Box flexDirection="column">
-        <Text>Which LangSmith region is this workspace in?</Text>
-        {LANGSMITH_REGION_OPTIONS.map((option, index) => (
-          <Text key={option.id}>
-            <SelectionMarker
-              isSelected={index === langsmithRegionSelectionIndex}
-            />{" "}
-            {option.name} <Text color="gray">({option.host})</Text>
-          </Text>
-        ))}
-        <Box flexDirection="column" marginTop={1}>
-          <Text color="gray">{selectedRegion.description}</Text>
-        </Box>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "global-cron-mode") {
-    return (
-      <Box flexDirection="column">
-        <Text>
-          {isCodeMode(onboardingConfig)
-            ? "When should GitHub Actions refresh this code wiki?"
-            : "When should OpenWiki run all ingestion?"}
-        </Text>
-        <Text color="gray">
-          {isCodeMode(onboardingConfig)
-            ? "OpenWiki will write a scheduled GitHub Actions workflow for this repository."
-            : "All configured sources run sequentially at this time."}
-        </Text>
-        <Text color="gray">Suggested: {suggestedCronDescription}</Text>
-        {CRON_MODE_OPTIONS.map((option, index) => (
-          <Text key={option}>
-            <SelectionMarker isSelected={index === cronModeSelectionIndex} />{" "}
-            {option}
-          </Text>
-        ))}
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "global-cron-custom") {
-    const validation = validateCronExpression(input);
-    return (
-      <Box flexDirection="column">
-        <Text>
-          {isCodeMode(onboardingConfig)
-            ? "Enter one GitHub Actions cron schedule for this code wiki."
-            : "Enter one cron schedule for all ingestion."}
-        </Text>
-        <SegmentedCronInput
-          activeFieldIndex={cronFieldSelectionIndex}
-          expression={input}
-          fallbackExpression={suggestedCronExpression}
-          maxDisplayWidth={inputDisplayWidth}
-        />
-        {input ? (
-          <Text color={validation.valid ? "cyan" : "red"}>
-            {validation.valid ? validation.description : validation.error}
-          </Text>
-        ) : (
-          <Text color="gray">Example: 0 2 * * *</Text>
-        )}
-        <Text color="gray">
-          Type in each field. Use right/left arrows or Tab to move; spaces also
-          move fields.
-        </Text>
-        <Text color="gray">Press Enter to save a valid schedule.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "global-power-mode") {
-    return (
-      <Box flexDirection="column">
-        <Text>Keep your Mac awake for scheduled refreshes?</Text>
-        <Text color="gray">
-          OpenWiki can use macOS pmset to wake 2 minutes before the shared
-          ingestion schedule and sleep 30 minutes after it.
-        </Text>
-        {sourceState.savedScheduleWarning ? (
-          <Text color="yellow">{sourceState.savedScheduleWarning}</Text>
-        ) : null}
-        <Box flexDirection="column" marginTop={1}>
-          {POWER_MODE_OPTIONS.map((option, index) => (
-            <Text key={option}>
-              <SelectionMarker isSelected={index === powerModeSelectionIndex} />{" "}
-              {option}
-            </Text>
-          ))}
-        </Box>
-        <Text color="gray">
-          macOS has one global repeat power schedule. Setting this can replace
-          an existing pmset repeat wake/sleep schedule.
-        </Text>
-      </Box>
-    );
-  }
-
-  if (step === "source-confirm-continue") {
-    const missingSources = sourceOptions.filter(
-      (source) => getSourceInstanceCount(onboardingConfig, source.id) === 0,
-    );
-    return (
-      <Box flexDirection="column">
-        <Text>Some sources for this mode are not configured yet.</Text>
-        {missingSources.map((source) => (
-          <Text color="gray" key={source.id}>
-            - {source.displayName}
-          </Text>
-        ))}
-        <Box flexDirection="column" marginTop={1}>
-          {SOURCE_CONTINUE_OPTIONS.map((option, index) => (
-            <Text key={option}>
-              <SelectionMarker
-                isSelected={index === sourceContinueSelectionIndex}
-              />{" "}
-              {option}
-            </Text>
-          ))}
-        </Box>
-        <Text color="gray">Use up/down arrows, then press Enter.</Text>
-      </Box>
-    );
-  }
-
-  if (step === "final") {
-    return (
-      <Box flexDirection="column">
-        <Text>Setup is complete.</Text>
-        {FINAL_OPTIONS.map((option, index) => {
-          const label = getFinalOptionLabel(option, selectedMode);
-          return (
-            <Text key={option}>
-              <SelectionMarker isSelected={index === finalSelectionIndex} />{" "}
-              {label}
-            </Text>
-          );
-        })}
-        <Text color="gray">
-          {selectedMode === "code"
-            ? "Run now writes the initial openwiki/ directory. Open chat skips the initial run."
-            : "Run now executes one source-specific ingestion and wiki update per configured source. Run later opens chat so you can start ingestion when you are ready."}
-        </Text>
-      </Box>
-    );
-  }
-
-  return null;
-}
-
-function mask(value: string): string {
-  if (value.length === 0) {
-    return "";
-  }
-
-  return "*".repeat(value.length);
-}
-
-function ExternalCliAuthPrompt({
-  authState,
-  input,
-  provider,
-}: {
-  authState: ExternalCliAuthState;
-  input: string;
-  provider: OpenWikiProvider;
-}) {
-  const adapter = getExternalCliAuthAdapter(provider);
-  const envKey = getProviderApiKeyEnvKey(provider) ?? "API key";
-
-  if (!adapter) {
-    return null;
-  }
-
-  if (authState.kind === "idle" || authState.kind === "checking") {
-    return (
-      <Text color="gray">
-        Checking for an existing {adapter.credentialDescription}...
-      </Text>
-    );
-  }
-
-  if (authState.kind === "logging-in") {
-    return (
-      <Text color="gray">
-        Running `{adapter.loginCommand}` — follow the prompts in this
-        terminal...
-      </Text>
-    );
-  }
-
-  if (authState.kind === "detected") {
-    return (
-      <Box flexDirection="column">
-        <Text>Detected an existing {adapter.credentialDescription}.</Text>
-        <Text color="gray">
-          Press Enter to use it, Tab to sign in again, or paste a different
-          token below.
-        </Text>
-        <Text>
-          <Text color="gray">$</Text> {envKey}={" "}
-          <Text color="yellow">
-            {input.length > 0 ? mask(input) : `<from ${adapter.name}>`}
-          </Text>
-        </Text>
-      </Box>
-    );
-  }
-
-  return (
-    <Box flexDirection="column">
-      <Text>No {adapter.credentialDescription} detected.</Text>
-      {authState.kind === "login-failed" ? (
-        <Text color="red">
-          `{adapter.loginCommand}` did not complete successfully.
-        </Text>
-      ) : null}
-      {authState.kind === "not-detected" && authState.cliAvailable ? (
-        <Text color="gray">
-          Press Tab to run `{adapter.loginCommand}`, or paste a token below.
-        </Text>
-      ) : (
-        <Text color="gray">
-          {adapter.installHint} You can also paste a token below for CI or other
-          headless use.
-        </Text>
-      )}
-      <Text>
-        <Text color="gray">$</Text> {envKey}={" "}
-        <Text color="yellow">{mask(input)}</Text>
-      </Text>
-      <Text color="gray">Press Enter to save it.</Text>
-    </Box>
-  );
-}
-
-function SetupHeader() {
-  return (
-    <Box
-      borderStyle="round"
-      borderColor="cyan"
-      flexDirection="column"
-      marginBottom={1}
-      paddingX={1}
-    >
-      <Text>
-        <Text bold color="cyan">
-          OpenWiki
-        </Text>{" "}
-        <Text color="gray">first-run setup</Text>
-      </Text>
-      <Text>Configure the model, wiki scope, and sources.</Text>
-    </Box>
-  );
-}
-
-type SetupStepState = "current" | "done" | "optional" | "pending";
-
-/**
- * Resolve a checklist row's status. The active step wins, so navigating back to
- * an already-done step shows the current-row cursor rather than a check; a done
- * step reads done; anything else falls to its resting status.
- */
-export function resolveStepStatus(
-  id: PromptStep,
-  activeStep: PromptStep | null,
-  done: boolean,
-  resting: "optional" | "pending" = "pending",
-): SetupStepState {
-  if (id === activeStep) {
-    return "current";
-  }
-  if (done) {
-    return "done";
-  }
-  return resting;
-}
-
-/**
- * Progress glyph per status: a check for done, an arrow for the active row, a
- * hollow circle for not-started (and optional). Single cell wide so every row's
- * label column lines up without padding the marker.
- */
-const STEP_GLYPH: Record<SetupStepState, string> = {
-  done: "✓",
-  current: "❯",
-  optional: "○",
-  pending: "○",
-};
-
-/** Color per status. Optionality is conveyed by the detail text, not the glyph. */
-const STEP_COLOR: Record<SetupStepState, string> = {
-  done: "green",
-  current: "cyan",
-  optional: "gray",
-  pending: "gray",
-};
-
-function SetupStep({
-  detail,
-  label,
-  state,
-}: {
-  detail: string;
-  label: string;
-  state: SetupStepState;
-}) {
-  return (
-    <Text>
-      <Text color={STEP_COLOR[state]}>{STEP_GLYPH[state]}</Text>{" "}
-      <Text bold={state === "current" || state === "done"}>
-        {label.padEnd(16)}
-      </Text>{" "}
-      <Text color="gray">{detail}</Text>
-    </Text>
-  );
-}
-
-function SetupPanel({
-  children,
-  title,
-}: {
-  children: React.ReactNode;
-  title: string;
-}) {
-  return (
-    <Box
-      borderStyle="single"
-      borderColor="gray"
-      flexDirection="column"
-      marginTop={1}
-      paddingX={1}
-    >
-      <Text bold color="cyan">
-        {title}
-      </Text>
-      {children}
-    </Box>
-  );
-}
-
-function SelectionMarker({ isSelected }: { isSelected: boolean }) {
-  return (
-    <Text color={isSelected ? "cyan" : "gray"}>{isSelected ? ">" : " "}</Text>
-  );
-}
-
-function SourceConnectionStatus({
-  count,
-  isConfigured,
-}: {
-  count: number;
-  isConfigured: boolean;
-}) {
-  return (
-    <Text color={isConfigured ? "green" : "gray"}>
-      {isConfigured
-        ? `[configured${count > 1 ? ` x${count}` : ""}]`
-        : "[not configured]"}
-    </Text>
-  );
-}
-
-function OAuthAuthorizationLink({
-  authProvider,
-  copiedToClipboard,
-  url,
-}: {
-  authProvider?: AuthProviderId;
-  copiedToClipboard: boolean;
-  url: string;
-}) {
-  return (
-    <Box flexDirection="column" marginTop={1}>
-      <Text>
-        <Text color="cyan" underline>
-          {formatTerminalHyperlink(url, "Open authorization URL")}
-        </Text>
-      </Text>
-      <Text color={copiedToClipboard ? "green" : "gray"}>
-        {getOAuthAuthorizationStatusText({
-          authProvider,
-          copiedToClipboard,
-        })}
-      </Text>
-    </Box>
-  );
-}
-
-export function getOAuthAuthorizationStatusText({
-  authProvider,
-  copiedToClipboard,
-}: {
-  authProvider?: AuthProviderId;
-  copiedToClipboard: boolean;
-}): string {
-  if (copiedToClipboard) {
-    return "Full URL copied to clipboard. Use the link above if your terminal supports it.";
-  }
-
-  const authCommand = authProvider
-    ? `openwiki auth ${authProvider}`
-    : "openwiki auth <provider>";
-
-  return `Use the terminal link above. If it is not clickable, cancel and run ${authCommand} in a plain terminal.`;
-}
-
-function OAuthLoginPrompt({
-  copied,
-  input,
-  isLoggingIn,
-  loginUrl,
-  provider,
-}: {
-  copied: boolean;
-  input: string;
-  isLoggingIn: boolean;
-  loginUrl: string | null;
-  provider: OpenWikiProvider;
-}) {
-  return (
-    <Box flexDirection="column" marginBottom={1}>
-      <Text bold color="cyan">
-        ChatGPT login
-      </Text>
-      <Text>
-        Sign in with your {getProviderLabel(provider)} account to authorize
-        OpenWiki.
-      </Text>
-      {loginUrl ? (
-        <Box flexDirection="column" marginTop={1}>
-          <Text color="gray">
-            Opening your browser. If it does not open, copy this URL:
-          </Text>
-          <Text color="cyan" wrap="wrap">
-            {loginUrl}
-          </Text>
-          <Text color="gray">
-            Press <Text bold>c</Text> to copy the URL
-            {copied ? <Text color="green"> (copied)</Text> : null}
-          </Text>
-          <Box flexDirection="column" marginTop={1}>
-            <Text color="gray">
-              If the browser cannot reach this machine, paste the redirect URL
-              or authorization code and press Enter:
-            </Text>
-            <Text>
-              <Text color="gray">&gt; </Text>
-              {input.length > 0 ? (
-                <Text color="yellow">{input}</Text>
-              ) : (
-                <Text color="gray">(paste here)</Text>
-              )}
-            </Text>
-          </Box>
-        </Box>
-      ) : (
-        <Text color="gray">Starting the ChatGPT login...</Text>
-      )}
-      <Text color="gray">
-        {isLoggingIn
-          ? "Waiting for browser sign-in or pasted URL..."
-          : "Login failed. Press Enter to retry."}
-      </Text>
-    </Box>
-  );
-}
-
-function BorderedInput({
-  borderColor = "cyan",
-  maxDisplayWidth,
-  marginTop,
-  prefix,
-  secret = false,
-  showCursor = true,
-  value,
-}: {
-  borderColor?: "cyan" | "gray";
-  maxDisplayWidth: number;
-  marginTop?: number;
-  prefix?: string;
-  secret?: boolean;
-  showCursor?: boolean;
-  value: string;
-}) {
-  const prompt = prefix ? "$ " : "> ";
-  const prefixText = prefix ? `${prefix} ` : "";
-  const valueDisplayWidth = Math.max(
-    1,
-    maxDisplayWidth - prompt.length - prefixText.length - (showCursor ? 1 : 0),
-  );
-
-  return (
-    <Box
-      borderStyle="single"
-      borderColor={borderColor}
-      marginTop={marginTop}
-      paddingX={1}
-      width={maxDisplayWidth + 4}
-    >
-      <Text wrap="truncate">
-        <Text color="gray">{prompt}</Text>
-        {prefixText ? <Text color="gray">{prefixText}</Text> : null}
-        <InputValueWithCursor
-          maxDisplayWidth={valueDisplayWidth}
-          secret={secret}
-          showCursor={showCursor}
-          value={value}
-        />
-      </Text>
-    </Box>
-  );
-}
-
-function BorderedMultilineInput({
-  borderColor = "cyan",
-  maxDisplayWidth,
-  marginTop,
-  showCursor = true,
-  value,
-}: {
-  borderColor?: "cyan" | "gray";
-  maxDisplayWidth: number;
-  marginTop?: number;
-  showCursor?: boolean;
-  value: string;
-}) {
-  return (
-    <Box
-      borderStyle="single"
-      borderColor={borderColor}
-      flexDirection="column"
-      marginTop={marginTop}
-      paddingX={1}
-      width={maxDisplayWidth + 4}
-    >
-      <Text wrap="wrap">
-        <Text color="gray">&gt; </Text>
-        {value ? <Text color="yellow">{value}</Text> : null}
-        {showCursor ? <Text inverse> </Text> : null}
-      </Text>
-    </Box>
-  );
-}
-
-function InputValueWithCursor({
-  maxDisplayWidth,
-  secret = false,
-  showCursor = true,
-  value,
-}: {
-  maxDisplayWidth: number;
-  secret?: boolean;
-  showCursor?: boolean;
-  value: string;
-}) {
-  if (secret) {
-    const displayValue = getSingleLineInputDisplayValue(
-      formatSecretInputDisplay(value),
-      maxDisplayWidth,
-    );
-
-    return (
-      <>
-        <Text color={value.length > 0 ? "yellow" : "gray"}>{displayValue}</Text>
-        {showCursor ? <Text inverse> </Text> : null}
-      </>
-    );
-  }
-
-  const displayValue = getSingleLineInputDisplayValue(value, maxDisplayWidth);
-
-  return (
-    <>
-      {displayValue ? <Text color="yellow">{displayValue}</Text> : null}
-      {showCursor ? <Text inverse> </Text> : null}
-    </>
-  );
-}
-
-function formatSecretInputDisplay(value: string): string {
-  // Empty renders as nothing (just the cursor); dots for the entered length,
-  // matching the non-secret inputs rather than printing a literal "empty".
-  return "•".repeat(value.length);
-}
-
-function formatTerminalHyperlink(url: string, label: string): string {
-  return `\u001B]8;;${url}\u0007${label}\u001B]8;;\u0007`;
-}
-
-function getSingleLineInputDisplayValue(
-  value: string,
-  maxLength: number,
-): string {
-  if (maxLength <= 0) {
-    return "";
-  }
-
-  if (value.length <= maxLength) {
-    return value;
-  }
-
-  if (maxLength <= 3) {
-    return value.slice(-maxLength);
-  }
-
-  return `...${value.slice(-(maxLength - 3))}`;
-}
-
-function SegmentedCronInput({
-  activeFieldIndex,
-  expression,
-  fallbackExpression,
-  maxDisplayWidth,
-}: {
-  activeFieldIndex: number;
-  expression: string;
-  fallbackExpression: string;
-  maxDisplayWidth: number;
-}) {
-  const fields = getCronFields(expression, fallbackExpression);
-  const fieldDisplayWidth = Math.max(
-    8,
-    Math.min(14, Math.floor(maxDisplayWidth / CRON_FIELD_LABELS.length) - 1),
-  );
-
-  return (
-    <Box flexDirection="column" marginTop={1}>
-      <Box>
-        {fields.map((field, index) => (
-          <Box
-            flexDirection="column"
-            marginRight={1}
-            key={CRON_FIELD_LABELS[index]}
-          >
-            <Text color="gray">{CRON_FIELD_LABELS[index]}</Text>
-            <BorderedInput
-              borderColor={index === activeFieldIndex ? "cyan" : "gray"}
-              maxDisplayWidth={fieldDisplayWidth}
-              showCursor={index === activeFieldIndex}
-              value={field}
-            />
-          </Box>
-        ))}
-      </Box>
-      <Text color="gray">Cron: {fields.join(" ")}</Text>
-    </Box>
-  );
-}
-
-export function getInitialStep(
-  modelIdOverride: string | null,
-  provider: OpenWikiProvider,
-  onboardingConfig: OpenWikiOnboardingConfig = createEmptyOnboardingConfig(),
-  mode: OpenWikiRunMode = "code",
-  allowModeSelection = false,
-  walkAll = false,
-): PromptStep | null {
-  if (walkAll) {
-    // Explicit --init: always start at the top and walk every applicable step,
-    // even ones already configured, instead of skipping to the first unset one.
-    return orderedSetupSteps(provider, mode, allowModeSelection)[0] ?? null;
-  }
-
-  if (allowModeSelection) {
-    return "run-mode";
-  }
-
-  if (!hasValidConfiguredProvider()) {
-    return "provider";
-  }
-
-  if (needsAwsCredentialRepair(provider)) {
-    return "region";
-  }
-
-  const nextCredentialStep = credentialStep(provider);
-
-  if (needsCredentialStep(provider) && nextCredentialStep) {
-    return nextCredentialStep;
-  }
-
-  if (needsSecretKeyStep(provider)) {
-    return "secret-key";
-  }
-
-  if (needsGcpProjectStep(provider)) {
-    return "gcp-project";
-  }
-
-  if (needsBaseUrlStep(provider)) {
-    return "base-url";
-  }
-
-  if (needsRegionStep(provider)) {
-    return "region";
-  }
-
-  if (
-    modelIdOverride === null &&
-    process.env[OPENWIKI_MODEL_ID_ENV_KEY] === undefined
-  ) {
-    return "model";
-  }
-
-  if (!process.env.LANGSMITH_API_KEY) {
-    return "langsmith";
-  }
-
-  if (mode === "code" && !isOnboardingComplete(onboardingConfig)) {
-    return "code-repo-confirm";
-  }
-
-  if (!getConfigModeId(onboardingConfig)) {
-    return "template";
-  }
-
-  if (!onboardingConfig.wikiGoal) {
-    return "wiki-goal";
-  }
-
-  if (!isCodeMode(onboardingConfig) && !onboardingConfig.ingestionSchedule) {
-    return "global-cron-mode";
-  }
-
-  if (!isOnboardingComplete(onboardingConfig)) {
-    return "source-menu";
-  }
-
-  return null;
-}
-
-export function getNextStepAfterProvider(
-  provider: OpenWikiProvider,
-  modelIdOverride: string | null,
-  onboardingConfig: OpenWikiOnboardingConfig = createEmptyOnboardingConfig(),
-  mode: OpenWikiRunMode = "code",
-  forceModelStep = false,
-): PromptStep | null {
-  if (needsAwsCredentialRepair(provider)) {
-    return "region";
-  }
-
-  const nextCredentialStep = credentialStep(provider);
-
-  if (needsCredentialStep(provider) && nextCredentialStep) {
-    return nextCredentialStep;
-  }
-
-  return getNextStepAfterApiKey(
-    provider,
-    modelIdOverride,
-    onboardingConfig,
-    mode,
-    forceModelStep,
-  );
-}
-
-function getNextStepAfterApiKey(
-  provider: OpenWikiProvider,
-  modelIdOverride: string | null,
-  onboardingConfig: OpenWikiOnboardingConfig,
-  mode: OpenWikiRunMode,
-  forceModelStep = false,
-): PromptStep | null {
-  if (needsSecretKeyStep(provider)) {
-    return "secret-key";
-  }
-
-  return getNextStepAfterSecretKey(
-    provider,
-    modelIdOverride,
-    onboardingConfig,
-    mode,
-    forceModelStep,
-  );
-}
-
-function getNextStepAfterSecretKey(
-  provider: OpenWikiProvider,
-  modelIdOverride: string | null,
-  onboardingConfig: OpenWikiOnboardingConfig,
-  mode: OpenWikiRunMode,
-  forceModelStep = false,
-): PromptStep | null {
-  if (needsGcpProjectStep(provider)) {
-    return "gcp-project";
-  }
-
-  return getNextStepAfterGcpLocation(
-    provider,
-    modelIdOverride,
-    onboardingConfig,
-    mode,
-    forceModelStep,
-  );
-}
-
-function getNextStepAfterGcpLocation(
-  provider: OpenWikiProvider,
-  modelIdOverride: string | null,
-  onboardingConfig: OpenWikiOnboardingConfig = createEmptyOnboardingConfig(),
-  mode: OpenWikiRunMode = "code",
-  forceModelStep = false,
-): PromptStep | null {
-  if (needsBaseUrlStep(provider)) {
-    return "base-url";
-  }
-
-  return getNextStepAfterBaseUrl(
-    provider,
-    modelIdOverride,
-    onboardingConfig,
-    mode,
-    forceModelStep,
-  );
-}
-
-function getNextStepAfterBaseUrl(
-  provider: OpenWikiProvider,
-  modelIdOverride: string | null,
-  onboardingConfig: OpenWikiOnboardingConfig,
-  mode: OpenWikiRunMode,
-  forceModelStep = false,
-): PromptStep | null {
-  if (needsRegionStep(provider)) {
-    return "region";
-  }
-
-  return getNextStepAfterRegion(
-    provider,
-    modelIdOverride,
-    onboardingConfig,
-    mode,
-    forceModelStep,
-  );
-}
-
-function getNextStepAfterRegion(
-  provider: OpenWikiProvider,
-  modelIdOverride: string | null,
-  onboardingConfig: OpenWikiOnboardingConfig,
-  mode: OpenWikiRunMode,
-  forceModelStep = false,
-): PromptStep | null {
-  if (
-    modelIdOverride === null &&
-    (forceModelStep || process.env[OPENWIKI_MODEL_ID_ENV_KEY] === undefined)
-  ) {
-    return "model";
-  }
-
-  if (!process.env.LANGSMITH_API_KEY) {
-    return "langsmith";
-  }
-
-  if (mode === "code" && !isOnboardingComplete(onboardingConfig)) {
-    return "code-repo-confirm";
-  }
-
-  if (!getConfigModeId(onboardingConfig)) {
-    return "template";
-  }
-
-  if (!onboardingConfig.wikiGoal) {
-    return "wiki-goal";
-  }
-
-  if (!isCodeMode(onboardingConfig) && !onboardingConfig.ingestionSchedule) {
-    return "global-cron-mode";
-  }
-
-  if (!isOnboardingComplete(onboardingConfig)) {
-    return "source-menu";
-  }
-
-  return null;
-}
-
-export function ensureRunModeConfig(
-  config: OpenWikiOnboardingConfig,
-  mode: OpenWikiRunMode,
-): OpenWikiOnboardingConfig {
-  if (getConfigModeId(config) === mode) {
-    return mode === "code" && config.wikiGoal !== undefined
-      ? { ...config, wikiGoal: undefined }
-      : config;
-  }
-
-  const runModeTemplate = ONBOARDING_TEMPLATES.find(
-    (option) => option.id === mode,
-  );
-  if (!runModeTemplate) {
-    return config;
-  }
-
-  return {
-    ...config,
-    modeId: runModeTemplate.id,
-    modeName: runModeTemplate.name,
-    templateId: runModeTemplate.id,
-    templateName: runModeTemplate.name,
-    ...(mode === "code" ? { wikiGoal: undefined } : {}),
-  };
-}
-
-export async function hydrateRunModeConfig(
-  config: OpenWikiOnboardingConfig,
-  mode: OpenWikiRunMode,
-  repoRoot: string,
-): Promise<OpenWikiOnboardingConfig> {
-  if (mode !== "code") {
-    return config;
-  }
-
-  const wikiGoal = await readRepositoryWikiInstructions(repoRoot);
-
-  return { ...config, wikiGoal };
-}
-
-function getRunModeSelectionIndex(mode: OpenWikiRunMode): number {
-  const index = RUN_MODE_OPTIONS.findIndex((option) => option.id === mode);
-  return index === -1 ? 0 : index;
-}
-
-function getLangsmithRegionSelectionIndex(region: LangSmithRegion): number {
-  const index = LANGSMITH_REGION_OPTIONS.findIndex(
-    (option) => option.id === region,
-  );
-  return index === -1 ? 0 : index;
-}
-
-function getLangsmithRegionLabel(region: LangSmithRegion): string {
-  const option = LANGSMITH_REGION_OPTIONS.find((item) => item.id === region);
-  return option ? `${option.name} (${option.host})` : region;
-}
-
-function getRunModeName(mode: OpenWikiRunMode): string {
-  return RUN_MODE_OPTIONS.find((option) => option.id === mode)?.name ?? mode;
-}
-
-function getSourceOption(sourceId: ConnectorId): SourceSetupOption {
-  return (
-    SOURCE_OPTIONS.find((source) => source.id === sourceId) ?? SOURCE_OPTIONS[0]
-  );
-}
-
-function getConfigModeId(config: OpenWikiOnboardingConfig): string | undefined {
-  return config.modeId ?? config.templateId;
-}
-
-function getConfigModeName(
-  config: OpenWikiOnboardingConfig,
-): string | undefined {
-  return config.modeName ?? config.templateName;
-}
-
-function isCodeMode(config: OpenWikiOnboardingConfig): boolean {
-  return getConfigModeId(config) === "code";
-}
-
-function needsEnvValue(secretInput: SourceSecretInput): boolean {
-  return !process.env[secretInput.envKey];
-}
-
-function addSourceInstanceConfig(
-  config: OpenWikiOnboardingConfig,
-  sourceInstance: OpenWikiOnboardingConfig["sourceInstances"][number],
-): OpenWikiOnboardingConfig {
-  const sourceInstances = [...config.sourceInstances, sourceInstance];
-  return {
-    ...config,
-    sourceInstances,
-    sources: deriveLegacySources(sourceInstances),
-  };
-}
-
-function deriveLegacySources(
-  sourceInstances: OpenWikiOnboardingConfig["sourceInstances"],
-): OpenWikiOnboardingConfig["sources"] {
-  const sources: OpenWikiOnboardingConfig["sources"] = {};
-
-  for (const sourceInstance of sourceInstances) {
-    if (!sources[sourceInstance.connectorId]) {
-      sources[sourceInstance.connectorId] = {
-        connectedAt: sourceInstance.connectedAt,
-        connectorConfig: sourceInstance.connectorConfig,
-        ingestionGoal: sourceInstance.ingestionGoal,
-      };
-    }
-  }
-
-  return sources;
-}
-
-function getSourceInstanceCount(
-  config: OpenWikiOnboardingConfig,
-  sourceId: ConnectorId,
-): number {
-  return getSourceInstances(config, sourceId).length;
-}
-
-function getSourceInstances(
-  config: OpenWikiOnboardingConfig,
-  sourceId: ConnectorId,
-): OpenWikiOnboardingConfig["sourceInstances"] {
-  return config.sourceInstances.filter(
-    (sourceInstance) => sourceInstance.connectorId === sourceId,
-  );
-}
-
-function getConnectedSourceCount(
-  config: OpenWikiOnboardingConfig,
-  sourceOptions: readonly SourceSetupOption[],
-): number {
-  const sourceIds = new Set(sourceOptions.map((source) => source.id));
-  return config.sourceInstances.filter((sourceInstance) =>
-    sourceIds.has(sourceInstance.connectorId),
-  ).length;
-}
-
-function createSourceInstanceId(
-  sourceId: ConnectorId,
-  config: OpenWikiOnboardingConfig,
-): string {
-  const sourceCount = getSourceInstanceCount(config, sourceId) + 1;
-  return `${sourceId}-${sourceCount}`;
-}
-
-function createSourceInstanceName(
-  source: SourceSetupOption,
-  description: string,
-  config: OpenWikiOnboardingConfig,
-): string {
-  const sourceCount = getSourceInstanceCount(config, source.id) + 1;
-  const trimmedDescription = description.trim();
-  const suffix = trimmedDescription.length > 0 ? `: ${trimmedDescription}` : "";
-  return `${source.displayName} ${sourceCount}${suffix}`.slice(0, 120);
-}
-
-function isSourceStep(step: PromptStep | null): boolean {
-  return Boolean(step?.startsWith("source-"));
-}
-
-function isScheduleStep(step: PromptStep | null): boolean {
-  return Boolean(step?.startsWith("global-"));
-}
-
-/**
- * Label for the provider's primary credential input. Bedrock authenticates
- * with an IAM access key ID (paired with a secret access key), not a single
- * opaque API key, so its prompt reads differently from every other provider.
- */
-function getApiKeyFieldLabel(provider: OpenWikiProvider): string {
-  return provider === "bedrock"
-    ? `${getProviderLabel(provider)} access key ID`
-    : `${getProviderLabel(provider)} API key`;
-}
-
-function hasValidConfiguredProvider(): boolean {
-  return normalizeProvider(process.env[OPENWIKI_PROVIDER_ENV_KEY]) !== null;
-}
-
-function getModelSetupDetail(
-  modelIdOverride: string | null,
-  provider: OpenWikiProvider,
-): string {
-  if (modelIdOverride) {
-    return `using ${modelIdOverride} for this run`;
-  }
-
-  if (process.env[OPENWIKI_MODEL_ID_ENV_KEY]) {
-    return process.env[OPENWIKI_MODEL_ID_ENV_KEY] ?? "";
-  }
-
-  return `default ${getDefaultModelId(provider)}`;
-}
-
-function getModelSelectionOptions(
-  provider: OpenWikiProvider,
-): ModelSelectionOption[] {
-  return [
-    ...getProviderModelOptions(provider).map((model) => ({
-      id: model.id,
-      kind: "preset" as const,
-      label: model.label,
-    })),
-    { kind: "custom" },
-  ];
-}
-
-function shouldStartWithCustomModelInput(provider: OpenWikiProvider): boolean {
-  return getProviderModelOptions(provider).length === 0;
-}
-
-function getSelectedModelId(
-  provider: OpenWikiProvider,
-  selectedIndex: number,
-  input: string,
-  isCustomInput: boolean,
-): string | null {
-  if (!isCustomInput) {
-    const selectedOption = getModelSelectionOptions(provider)[selectedIndex];
-
-    if (!selectedOption) {
-      return null;
-    }
-
-    return selectedOption.kind === "custom" ? "custom" : selectedOption.id;
-  }
-
-  const normalizedModelId = normalizeModelId(input);
-
-  return isValidModelId(normalizedModelId) ? normalizedModelId : null;
-}
-
-function getProviderSelectionIndex(provider: OpenWikiProvider): number {
-  const selectedIndex = SELECTABLE_OPENWIKI_PROVIDERS.findIndex(
-    (providerOption) => providerOption === provider,
-  );
-
-  return selectedIndex === -1 ? 0 : selectedIndex;
-}
-
-function getModelSelectionIndex(
-  provider: OpenWikiProvider,
-  selectedModelId: string,
-): number {
-  const selectedIndex = getModelSelectionOptions(provider).findIndex(
-    (option) => option.kind === "preset" && option.id === selectedModelId,
-  );
-
-  return selectedIndex === -1 ? 0 : selectedIndex;
-}
-
-function moveSelectionIndex(
-  currentIndex: number,
-  offset: number,
-  itemCount: number,
-): number {
-  if (itemCount <= 0) {
-    return 0;
-  }
-
-  return (currentIndex + offset + itemCount) % itemCount;
-}
-
-function getInputDisplayWidth(stdoutColumns: number | undefined): number {
-  const defaultWidth = 64;
-
-  if (!stdoutColumns || stdoutColumns <= 0) {
-    return defaultWidth;
-  }
-
-  return Math.max(24, Math.min(96, stdoutColumns - 16));
-}
-
-function getProviderArticle(provider: OpenWikiProvider): "a" | "an" {
-  return provider === "baseten" ||
-    provider === "fireworks" ||
-    provider === "gemini" ||
-    provider === "gemini-enterprise" ||
-    provider === "nebius"
-    ? "a"
-    : "an";
-}
-
-function getTemplateGoal(templateId: string | undefined): string {
-  return (
-    ONBOARDING_TEMPLATES.find((template) => template.id === templateId)
-      ?.suggestedGoal ?? ""
-  );
-}
-
-function getSourceMenuLabel(
-  source: SourceSetupOption,
-  sourceInstanceCount: number,
-): string {
-  return sourceInstanceCount > 0
-    ? `Add another ${source.displayName}`
-    : `Add ${source.displayName}`;
-}
-
-function getTemplateSourceOptions(
-  templateId: string | undefined,
-): readonly SourceSetupOption[] {
-  const template =
-    ONBOARDING_TEMPLATES.find((option) => option.id === templateId) ??
-    ONBOARDING_TEMPLATES[0];
-  const sourceIds = new Set(template.sourceIds);
-  const sourceOptions = SOURCE_OPTIONS.filter((source) =>
-    sourceIds.has(source.id),
-  );
-
-  return sourceOptions.length > 0 ? sourceOptions : SOURCE_OPTIONS;
-}
-
-function getSourceDescriptionPrompt(source: SourceSetupOption): string {
-  if (source.id === "web-search") {
-    return "Describe the topics, companies, or pages OpenWiki should search for.";
-  }
-
-  if (source.id === "hackernews") {
-    return "Describe the topics, keywords, users, or story types OpenWiki should watch on Hacker News.";
-  }
-
-  if (source.id === "git-repo") {
-    return "Describe what OpenWiki should understand about this repository.";
-  }
-
-  return `Describe what OpenWiki should look for in ${source.displayName}.`;
-}
-
-function getFinalOptionLabel(
-  option: (typeof FINAL_OPTIONS)[number],
-  mode: OpenWikiRunMode,
-): string {
-  if (mode !== "code") {
-    return option;
-  }
-
-  return option === "Run ingestion now" ? "Run OpenWiki now" : "Open chat";
-}
-
-function getSourceDescriptionOptionCount(source: SourceSetupOption): number {
-  return source.examples.length + 1;
-}
-
-function handleCronEditorInput({
-  currentFieldIndex,
-  currentValue,
-  fallbackExpression,
-  inputValue,
-  key,
-  replaceCurrentField,
-  setCurrentFieldIndex,
-  setReplaceCurrentField,
-  setValue,
-}: {
-  currentFieldIndex: number;
-  currentValue: string;
-  fallbackExpression: string;
-  inputValue: string;
-  key: PromptInputKey;
-  replaceCurrentField: boolean;
-  setCurrentFieldIndex: React.Dispatch<React.SetStateAction<number>>;
-  setReplaceCurrentField: React.Dispatch<React.SetStateAction<boolean>>;
-  setValue: React.Dispatch<React.SetStateAction<string>>;
-}): boolean {
-  if (key.leftArrow) {
-    setCurrentFieldIndex((index) => Math.max(0, index - 1));
-    setReplaceCurrentField(true);
-    return true;
-  }
-
-  if (key.rightArrow || key.tab || inputValue === " " || inputValue === "\t") {
-    setCurrentFieldIndex((index) =>
-      Math.min(CRON_FIELD_LABELS.length - 1, index + 1),
-    );
-    setReplaceCurrentField(true);
-    return true;
-  }
-
-  if (key.backspace || key.delete) {
-    const fields = getCronFields(currentValue, fallbackExpression);
-    const currentField = fields[currentFieldIndex] ?? "";
-    if (currentField.length === 0 && currentFieldIndex > 0) {
-      setCurrentFieldIndex(currentFieldIndex - 1);
-      setReplaceCurrentField(false);
-      return true;
-    }
-
-    fields[currentFieldIndex] = currentField.slice(0, -1);
-    setValue(fields.join(" "));
-    setReplaceCurrentField(false);
-    return true;
-  }
-
-  if (key.ctrl || key.meta) {
-    return false;
-  }
-
-  const pastedFields = parseCronFieldPaste(inputValue);
-  if (pastedFields.length > 1) {
-    const fields = getCronFields(currentValue, fallbackExpression);
-    pastedFields.forEach((field, offset) => {
-      const fieldIndex = currentFieldIndex + offset;
-      if (fieldIndex < CRON_FIELD_LABELS.length) {
-        fields[fieldIndex] = field;
-      }
-    });
-    setValue(fields.join(" "));
-    setCurrentFieldIndex((index) =>
-      Math.min(CRON_FIELD_LABELS.length - 1, index + pastedFields.length - 1),
-    );
-    setReplaceCurrentField(true);
-    return true;
-  }
-
-  const sanitizedInput = sanitizeCronInputChunk(inputValue);
-
-  if (!sanitizedInput) {
-    return false;
-  }
-
-  const fields = getCronFields(currentValue, fallbackExpression);
-  fields[currentFieldIndex] = replaceCurrentField
-    ? sanitizedInput
-    : `${fields[currentFieldIndex] ?? ""}${sanitizedInput}`;
-  setValue(fields.join(" "));
-  setReplaceCurrentField(false);
-  return true;
-}
-
-function getCronFields(
-  expression: string,
-  fallbackExpression: string,
-): string[] {
-  const source =
-    expression.trim().length > 0 ? expression.trim() : fallbackExpression;
-  const fields = source.split(/\s+/u);
-
-  return CRON_FIELD_LABELS.map((_, index) => fields[index] ?? "");
-}
-
-function parseCronFieldPaste(inputValue: string): string[] {
-  if (inputValue.trim().length === 0) {
-    return [];
-  }
-
-  if (/\s/u.test(inputValue)) {
-    return inputValue
-      .trim()
-      .split(/\s+/u)
-      .map((field) => sanitizeCronInputChunk(field))
-      .filter((field) => field.length > 0);
-  }
-
-  const compactValue = sanitizeCronInputChunk(inputValue);
-
-  if (/^[0-9*]{5}$/u.test(compactValue)) {
-    return compactValue.split("");
-  }
-
-  return [];
-}
-
-function sanitizeInputChunk(value: string): string {
-  return value.replace(/[\r\n]/gu, "");
-}
-
-function sanitizeCronInputChunk(value: string): string {
-  return value.replace(/[^A-Za-z0-9*,/?#LW.-]/gu, "");
-}
-
-function sanitizeRepoId(value: string): string {
-  return value.replace(/[^A-Za-z0-9._-]/gu, "-").slice(0, 80) || "repo";
-}
-
-function getDefaultLocalGitRepoPath(): string {
-  return process.cwd();
-}
-
-function getDefaultCodeRepoRootPath(): string {
-  return findNearestGitRepoRoot(process.cwd()) ?? process.cwd();
-}
-
-export function findNearestGitRepoRoot(startPath: string): string | null {
-  let currentPath = path.resolve(startPath);
-
-  while (true) {
-    if (existsSync(path.join(currentPath, ".git"))) {
-      return currentPath;
-    }
-
-    const parentPath = path.dirname(currentPath);
-    if (parentPath === currentPath) {
-      return null;
-    }
-
-    currentPath = parentPath;
-  }
-}
-
-async function validateLocalDirectoryPath(value: string): Promise<string> {
-  const normalizedPath = normalizeLocalPath(value);
-
-  if (normalizedPath.length === 0) {
-    throw new Error("Enter a local directory.");
-  }
-
-  const { stat } = await import("node:fs/promises");
-  const pathStat = await stat(normalizedPath);
-
-  if (!pathStat.isDirectory()) {
-    throw new Error(`${normalizedPath} is not a directory.`);
-  }
-
-  return normalizedPath;
-}
-
-function normalizeLocalPath(value: string): string {
-  const trimmedValue = value.trim();
-  if (trimmedValue.length === 0) {
-    return "";
-  }
-
-  if (trimmedValue === "~") {
-    return homedir();
-  }
-
-  if (trimmedValue.startsWith("~/") || trimmedValue.startsWith("~\\")) {
-    return path.resolve(homedir(), trimmedValue.slice(2));
-  }
-
-  return path.resolve(trimmedValue);
-}
-
-function getStaticSourceConfig(
-  sourceId: ConnectorId,
-  query: string,
-): Record<string, unknown> {
-  const queries = query.trim().length > 0 ? [query.trim()] : [];
-
-  if (sourceId === "web-search") {
-    return {
-      enabled: true,
-      includeAnswer: true,
-      includeImages: false,
-      includeRawContent: false,
-      maxResults: 5,
-      queries,
-      searchDepth: "basic",
-      timeRange: "day",
-      topic: "general",
-    };
-  }
-
-  if (sourceId === "hackernews") {
-    return {
-      enabled: true,
-      feeds: ["top", "new"],
-      maxItemsPerFeed: 30,
-      maxResultsPerQuery: 20,
-      queries,
-      queryTags: ["story"],
-    };
-  }
-
-  return {
-    enabled: true,
-  };
 }
 
 function getErrorMessage(error: unknown): string {
