+++ src/setup/inputs/cron.ts
@@ -1,0 +1,132 @@
+import type React from "react";
+import type { PromptInputKey } from "../types.js";
+
+/** Strips CR/LF from a pasted single-line input value. */
+export function sanitizeInputChunk(value: string): string {
+  return value.replace(/[\r\n]/gu, "");
+}
+
+export const CRON_FIELD_LABELS = ["minute", "hour", "day", "month", "weekday"];
+
+export function getCronFields(
+  expression: string,
+  fallbackExpression: string,
+): string[] {
+  const source =
+    expression.trim().length > 0 ? expression.trim() : fallbackExpression;
+  const fields = source.split(/\s+/u);
+
+  return CRON_FIELD_LABELS.map((_, index) => fields[index] ?? "");
+}
+
+export function parseCronFieldPaste(inputValue: string): string[] {
+  if (inputValue.trim().length === 0) {
+    return [];
+  }
+
+  if (/\s/u.test(inputValue)) {
+    return inputValue
+      .trim()
+      .split(/\s+/u)
+      .map((field) => sanitizeCronInputChunk(field))
+      .filter((field) => field.length > 0);
+  }
+
+  const compactValue = sanitizeCronInputChunk(inputValue);
+
+  if (/^[0-9*]{5}$/u.test(compactValue)) {
+    return compactValue.split("");
+  }
+
+  return [];
+}
+
+export function sanitizeCronInputChunk(value: string): string {
+  return value.replace(/[^A-Za-z0-9*,/?#LW.-]/gu, "");
+}
+
+export function handleCronEditorInput({
+  currentFieldIndex,
+  currentValue,
+  fallbackExpression,
+  inputValue,
+  key,
+  replaceCurrentField,
+  setCurrentFieldIndex,
+  setReplaceCurrentField,
+  setValue,
+}: {
+  currentFieldIndex: number;
+  currentValue: string;
+  fallbackExpression: string;
+  inputValue: string;
+  key: PromptInputKey;
+  replaceCurrentField: boolean;
+  setCurrentFieldIndex: React.Dispatch<React.SetStateAction<number>>;
+  setReplaceCurrentField: React.Dispatch<React.SetStateAction<boolean>>;
+  setValue: React.Dispatch<React.SetStateAction<string>>;
+}): boolean {
+  if (key.leftArrow) {
+    setCurrentFieldIndex((index) => Math.max(0, index - 1));
+    setReplaceCurrentField(true);
+    return true;
+  }
+
+  if (key.rightArrow || key.tab || inputValue === " " || inputValue === "\t") {
+    setCurrentFieldIndex((index) =>
+      Math.min(CRON_FIELD_LABELS.length - 1, index + 1),
+    );
+    setReplaceCurrentField(true);
+    return true;
+  }
+
+  if (key.backspace || key.delete) {
+    const fields = getCronFields(currentValue, fallbackExpression);
+    const currentField = fields[currentFieldIndex] ?? "";
+    if (currentField.length === 0 && currentFieldIndex > 0) {
+      setCurrentFieldIndex(currentFieldIndex - 1);
+      setReplaceCurrentField(false);
+      return true;
+    }
+
+    fields[currentFieldIndex] = currentField.slice(0, -1);
+    setValue(fields.join(" "));
+    setReplaceCurrentField(false);
+    return true;
+  }
+
+  if (key.ctrl || key.meta) {
+    return false;
+  }
+
+  const pastedFields = parseCronFieldPaste(inputValue);
+  if (pastedFields.length > 1) {
+    const fields = getCronFields(currentValue, fallbackExpression);
+    pastedFields.forEach((field, offset) => {
+      const fieldIndex = currentFieldIndex + offset;
+      if (fieldIndex < CRON_FIELD_LABELS.length) {
+        fields[fieldIndex] = field;
+      }
+    });
+    setValue(fields.join(" "));
+    setCurrentFieldIndex((index) =>
+      Math.min(CRON_FIELD_LABELS.length - 1, index + pastedFields.length - 1),
+    );
+    setReplaceCurrentField(true);
+    return true;
+  }
+
+  const sanitizedInput = sanitizeCronInputChunk(inputValue);
+
+  if (!sanitizedInput) {
+    return false;
+  }
+
+  const fields = getCronFields(currentValue, fallbackExpression);
+  fields[currentFieldIndex] = replaceCurrentField
+    ? sanitizedInput
+    : `${fields[currentFieldIndex] ?? ""}${sanitizedInput}`;
+  setValue(fields.join(" "));
+  setReplaceCurrentField(false);
+  return true;
+}
